Fix 64-bit compile errors
Cast to type appropriate for the print size.

+++ fabrics.c
@@ -269,7 +269,7 @@ static void print_discovery_log(struct nvmf_disc_rsp_page_hdr *log, int numrec)
 	int i;
 
 	printf("Discovery Log Number of Records %d, Generation counter %"PRIu64"\n",
-		numrec, __le64_to_cpu(log->genctr));
+		numrec, (uint64_t)__le64_to_cpu(log->genctr));
 
 	for (i = 0; i < numrec; i++) {
 		struct nvmf_disc_rsp_page_entry *e = &log->entries[i];
